uploading and using pics, chrono section, more macro
added pictures for the new hitbox system in GAS for the attack section. split the chronostasis mechanic away from the macro gameplay loop section and into its own to prevent a lot of bloat talking about a bunch of different things. and continued typing things out for the macro gameplay loop and a little for the new chrono section